Changed Add Product, Added Product fetching functions, Added Browsing functions, Added little bit to homepage
Reworked Add Product so that it now also includes the URL of a product as a field in its document. Created a function to fetch all product data from the firebase realtime database. Created a function to display products in grids when given an array of type Product. Created the home page with a select category menu and a search button to display all products or products of a certain category.