Delete Orders and Order Status
Added my delete orders function to the delete button on the orders page. It allows the user to now delete any of their orders. Orders that are created now will have a status variable that will be set to "processing" upon order confirmaton. I made a function to be able to change the order status of a certain order, still needs testing.